requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
black==19.3b0
certifi==2019.11.28
Click==7.0
colorama==0.4.1
Django==3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
wincertstore==0.2
wrapt==1.11.2
